// tests/regression/test/issue-2778/schema.ts
//////////////////////////////////////////////////////////////////////////////////////////////
// DO NOT MODIFY THIS FILE                                                                  //
// This file is automatically generated by ZenStack CLI and should not be manually updated. //
//////////////////////////////////////////////////////////////////////////////////////////////

/* eslint-disable */

import { type SchemaDef, type AttributeApplication, type FieldDefault, ExpressionUtils } from "@zenstackhq/schema";
export class SchemaType implements SchemaDef {
    provider = {
        type: "sqlite"
    } as const;
    models = {
        Tag: {
            name: "Tag",
            fields: {
                id: {
                    name: "id",
                    type: "String",
                    id: true,
                    attributes: [{ name: "@id" }, { name: "@default", args: [{ name: "value", value: ExpressionUtils.call("uuid") }] }] as readonly AttributeApplication[],
                    default: ExpressionUtils.call("uuid") as FieldDefault
                },
                name: {
                    name: "name",
                    type: "String",
                    unique: true,
                    attributes: [{ name: "@unique" }] as readonly AttributeApplication[]
                },
                products: {
                    name: "products",
                    type: "Product",
                    array: true,
                    relation: { opposite: "tags" }
                }
            },
            idFields: ["id"],
            uniqueFields: {
                id: { type: "String" },
                name: { type: "String" }
            }
        },
        Product: {
            name: "Product",
            fields: {
                id: {
                    name: "id",
                    type: "String",
                    id: true,
                    attributes: [{ name: "@id" }, { name: "@default", args: [{ name: "value", value: ExpressionUtils.call("uuid") }] }] as readonly AttributeApplication[],
                    default: ExpressionUtils.call("uuid") as FieldDefault
                },
                name: {
                    name: "name",
                    type: "String"
                },
                tags: {
                    name: "tags",
                    type: "Tag",
                    array: true,
                    relation: { opposite: "products" }
                }
            },
            idFields: ["id"],
            uniqueFields: {
                id: { type: "String" }
            }
        }
    } as const;
    plugins = {};
}
export const schema = new SchemaType();